backend/requirements.txt
ultralytics
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-telegram-bot
httpx
